WED 00:46 AM
Created an enemy spritesheet and update the confirm/summon button to be a spritesheet. Created an icon for how much blood the player has, found an icon for the amount of health. Created a spritesheet button to allow my confirm button to work.

Fixed an issue where spritesheets weren't aligned properly due to having set the margin property to 1 rather than spacing.
Added in text for how much health and how much blood the player has.
Created a temp tentacle that the player can drag around after pressing the summon button to choose where to place a tentacle. They then hit the confirm button to confirm the placement and a proper tentacle is added to the tentacle group.

Fixed an issue where the tentacle group had a normal phaser sprite in it. Added to the update function all the new items and it iterates through all the tentacles that have been summoned. Created a grid snapping function that rounds a value to the nearest multiple of 32.