WED 15:43
updated enemy base class to use the active property built into the sprite and added a callback scope to the reset timer that fixed the animation bug with the tentacles.

Created a new class WaveProperties to store the properties of each wave that you want to pass to your level to be store in a waves array. Started on making the game split up into waves and added several checks into the update loop to give different sections.